# src/main/resources/static/docs/trycatch_openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "trycatch API",
    "description" : "trycatch API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://trycatch"
  } ],
  "paths" : {
...
    "/<unknown>/roadmap/" : {
      "post" : {
        "summary" : "POST /v1/roadmap/",
        "operationId" : "registerRoadmap",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RoadmapRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/<unknown>/roadmap/{userName}" : {
      "get" : {
        "summary" : "GET /v1/roadmap/{userName}",
        "operationId" : "findRoadmap",
        "parameters" : [ {
          "name" : "userName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      },
      "put" : {
        "summary" : "PUT /v1/roadmap/{userName}",
        "operationId" : "modifyRoadmap",
        "parameters" : [ {
          "name" : "userName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RoadmapRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "DELETE /v1/roadmap/{userName}",
        "operationId" : "removeRoadmap",
        "parameters" : [ {
          "name" : "userName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "HttpHeaders" : {
        "type" : "object",
        "properties" : { }
      },
      "ResponseEntity" : {
        "type" : "object",
        "properties" : {
          "headers" : {
            "$ref" : "#/components/schemas/HttpHeaders"
          }
        }
      },
...
      "RoadmapRequestDto" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string"
          },
          "tag" : {
            "type" : "string"
          },
          "nodes" : {
            "type" : "string"
          },
          "edges" : {
            "type" : "string"
          }
        }
      }
    }
  }
}